protocol: Serialize handlePDUReceived()
Assigning a thread to each received PDU make COSMOS fail to reproduce
receiving file because Metadata PDU is not guarantied to be process at
the first.

COSMOS CFDP module assumes that a Metadata PDU is the first PUD for a
given transaction and create a transaction only when the module
received a Metadata.  if non- Metadata PUD is processed _before_
Metadata PDU creates a session, it throws away the given data PUD with
the error message "Received CFDP::PDUFileData of a non active class 1
transaction."

Because the module creates a thread for each receiving PDU, the
processing order isn't guarantied.  Thus, the module loose all PDUs
scheduled before Metadata PUD.

handlePDUReceived() already synchronizes with @transactionMutex anyway,
there is no need for threads.  This commit changes it to be
serialized.